pyproject.toml
[project]
name = "authserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "cryptography>=46.0.6",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "opentelemetry-distro>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39.0",
    "opentelemetry-instrumentation-asgi>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.61b0",
    "opentelemetry-instrumentation-psycopg2>=0.60b1",
    "opentelemetry-instrumentation-redis>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-instrumentation-system-metrics>=0.60b1",
    "opentelemetry-instrumentation-threading>=0.60b1",
    "opentelemetry-sdk>=1.39.0",
    "pika>=1.3.2",
    "prometheus-client>=0.24.1",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
